Extracted Cost class from Service & RoomCost.
