+++ src/main/java/seedu/address/model/hotel/bill/Service.java
@@ -11,27 +11,22 @@ public class Service implements Chargeable {
     public static final String DESC_MESSAGE_CONSTRAINTS =
             "Description should only contain alphanumeric characters and spaces, and it should not be blank";
     public static final String DESC_VALIDATION_REGEX = "[\\p{Alnum}]*";
-    public static final String COST_MESSAGE_CONSTRAINTS =
-            "Price should only contain numbers, and it should have a maximum of 2 decimal places";
-    public static final String COST_VALIDATION_REGEX = "[\\d\.\\\d{2}]";
 
     private String description;
-    private double cost;
+    private Cost cost;
 
     /**
-     * Creates a {@code service} that has the cost of {@code price}.
+     * Creates a {@code service} that has a {@code cost}.
      *
      * @param desc a valid description
      * @param cost a valid cost
      */
-    public Service(String desc, String cost) {
+    public Service(String desc, Cost cost) {
         requireNonNull(desc);
         checkArgument(isValidDescription(desc), DESC_MESSAGE_CONSTRAINTS);
-        requireNonNull(cost);
-        checkArgument(isValidCost(cost), COST_MESSAGE_CONSTRAINTS);
 
         this.description = desc;
-        this.cost = Double.valueOf(cost);
+        this.cost = cost;
     }
 
     /**
@@ -41,20 +36,12 @@ public class Service implements Chargeable {
         return test.matches(DESC_VALIDATION_REGEX);
     }
 
-    /**
-     * Returns true if a given string is a valid cost.
-     */
-    @Override
-    public static boolean isValidCost(String test) {
-        return test.matches(COST_VALIDATION_REGEX);
+    public String getDescription() {
+        return this.description;
     }
 
     @Override
     public double getCost() {
         return this.cost;
-    }
-
-    public String getDescription() {
-        return this.description;
     }
 }
